refactor internal pkg
